requirements.txt
datasets~=2.18.0
joblib~=1.3.2
matplotlib~=3.8.3
numpy~=1.26.4
pandas~=2.2.1
python-dotenv~=1.0.1
scikit-learn~=1.4.1.post1
torch~=2.2.1
tqdm~=4.66.2
transformers~=4.38.2
typer~=0.9.0
wandb~=0.16.6

rootutils~=1.0.7

plotly

jupyter
accelerate~=0.30.1

arize==7.11.1
arize-phoenix==3.16.1
arize-phoenix-evals==0.3.1

wordcloud==1.9.3